Backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.2

# LangChain core
langchain==0.0.350
langchain-core>=0.2.0

# Optional components
langchain-community==0.0.13
langchain-ollama==0.0.1

# Development tools
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# PDF processing
pypdf==3.17.1

# New dependencies
chromadb==0.4.22
faiss-cpu==1.7.4
pydantic-settings==2.1.0
pyyaml==6.0.1
python-dotenv==1.0.0
psutil==5.9.7
sqlparse>=0.4.4
sqlglot>=19.0.0
PyMuPDF==1.23.8